kv, roachpb: move seq num manipulation under Transaction
It belongs better there than as a standalone function, particularly
since we need to give special consideration to transaction fields (see
next commit).

Release note: None
